Test both with and without CGo
Since this would otherwise make our build matrix even larger, we drop
testing on Go "tip" for now. Functioning without CGo enabled is more
important than functioning on newer versions of Go.

With that said, we may eventually want to restore tip-testing for Windows
in particular, since our Windows implementation uses unsafe and so is
not subject to the Go 1.0 compatibility guarantee. We'll cross that bridge
when we come to it.